requirements.txt
aiogram==3.18.0
aiogram-dialog==2.3.1
tortoise-orm==0.24.2
ruff==0.10.0
envparse==0.2.0
emoji==2.14.1